+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] 2024-09-02
+### Added
+- Support for BetterBags Contexts
+
 ## [1.6.1] 2024-09-01
 ### Fixed
 - Fixed a typo in the configuration options.